+++ cmd/oras/internal/metadata/push.go
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package meta
+package metadata
 
 import ocispec "github.com/opencontainers/image-spec/specs-go/v1"
