+++ cmd/oras/internal/metadata/type.go
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package meta
+package metadata
 
 import (
 	"github.com/opencontainers/go-digest"
